Split StartController into specific Controllers
